basic profiling

+++ inverse_design/conditional_generator.py
@@ -34,7 +34,7 @@ def new_latent_design(shape, bias=0, r=None):
 # Cell
 @jax.jit
 def transform(latent, brush, beta=5.0):
-    convolved = conv2d(latent, jnp.asarray(brush, dtype=float)) / brush.sum()
+    convolved = conv2d(latent, jnp.asarray(brush, dtype=latent.dtype)) / brush.sum()
     return jnp.tanh(beta * convolved)
 
 # Cell
